Add files via upload
Welcome!
I created these two short compositions using Chuck programme language! 
It was not easy for me, at all...but I enjoyed it and at the moment I am satisfied with the outcome!
"Zensur" (Censorship) was based on rejected parts of "ego_echo" composition...and what I tried to do was to "play" with reverb and oscillators...in fact, I tried to control reverb through the change of oscillators in order to form a concrete musical form.
"Ego_echo" is indeed an egoistic sound ("I sound” in Greek) because it was my first attempt to construct a musical form using what I had understood from Chuck language! During its formation, I was thinking musically. That helped me a lot. It was not easy to make it sound but in the end, somehow it worked…